Ctrl-3 to put current window in center of left eye DK2 screen, better cascading
New hotkey: Ctrl-3

This will put the currently active window into center the left half of the first monitor, which should place it in front of the left eye on the Rift. Should help with hunting for windows on the Rift without having to take it off and move all windows.

Cascading improved:

The cascading of all moved windows on Ctrl-1 / -2 has been spaced wider so that windows are easier to find.